Correct message seqNo under lock
Since we construct the message with mutex not held, it might be constructed with seqNo not matching the sequence number of the current wal file.

We still want to construct the message outside of the lock.

Check if the seqNo is correct and if not reconstruct the message under lock and write the reconstructed message.